task/install: refrain from purging /var/lib/ceph
This is not necessary when the VMs are going to be destroyed, anyway.

Furthermore, the deepsea task has to bend over backwards to ensure
tests don't fail in this code. Not running it will simplify the
deepsea task.

Finally, purging /var/lib/ceph has little (or nothing) to do with the
install task.

Signed-off-by: Nathan Cutler <ncutler@suse.com>