Fix ipv4 and ipv6 address logging for Remote.run
The Remote class does not respect ip addresses
when it comes to define shortnames. As a result,
the hostname is not shown correctly in the log.
For ipv4 it only shows first number of the octet.
For ipv6 it even does not allow to proceed,
and raises exception in orchestra.run.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.de>
(cherry picked from commit fe38df6fce78ce26856ed148c803505f22bcc920)